+++ python_service/db/collections.py
@@ -1,0 +1,3 @@
+from .connection import db
+
+articles_collection = db["articles"]
