+++ python_service/routers/groups.py
@@ -1,0 +1,7 @@
+from fastapi import APIRouter
+
+router = APIRouter()
+
+@router.get("/")
+def get_articles_group():
+    return {"message": "This endpoint will return grouped articles."}
